Alpha, gamma and beta-components for taxonomic, functional and phylogenetic diversity

The Rao function computes alpha, gamma and beta-components for taxonomic, functional and phylogenetic diversity with the Rao index. The script integrates two functions: "Qdecomp", by Villeger et Mouillot (J Ecol, 2008) modified by Wilfried Thuiller, and "disc", by S. Pavoine, in the package ade4. For a regional assemblage of C local communities gamma = mean(alpha) + beta, where: gamma is the diversity of the regional pool, alpha is the diversity of the local community and beta is the turn over between local communities diversity is estimated with the Rao quadratic entropy index (Rao 1982)

Usage

RaoRel(sample, dfunc, dphyl, weight = FALSE, Jost = FALSE, 
  structure = NULL)

Arguments

sample

Community matrix of abundance (c x s) of the s species for the c local communities.

dfunc

matrix (s x s) or dist object with pairwise functional trait distances between the s species

dphyl

As dfunct but for phylogenetic distances

weight

Defining if the correction by Villeger & Mouillot (J Ecol, 2008) is applied or not

Jost

Defining if the Jost correction is applied (Jost 2007)

structure

A data frame containing the name of the group to which samples belong see de Bello et al, 2011 for more details.

Details

NA are automatically replaced by 0 in "sample". This function use the function "Qdecomp" by Sebastien Villeger & David Mouillot (J Ecol, 2008) modified by Wilfried Thuiller and the function disc originally proposed by Sandrine Pavoine.

Value

The results are organized for Taxonomic diversity ($TD), Functional diversity ($FD) and phylogenetical diversity ($PD). Beta and gamma diversities are calculated for the whole data set and for each pair of samples ("Pairwise_samples"):

$Richness_per_plot(number of species per sample)

$Relative_abundance (species relative abundances per plot)

$Pi (species regional relative abundance)

$Wc (weigthing factor),

$Mean_Alpha (mean aplpha diversity; for taxonomic diversity the Simpson index is calculated)

$Alpha (alpha diversity for each sample; for taxonomic diversity the Simpson index is calculated)

$Gamma (gamma diversity; for taxonomic diversity the Simpson index is calculated)

$Beta_add (Gamma-Mean_Alpha)

$Beta_prop (Beta_add*100/Gamma)

$Pairwise_samples$Alpha (mean alpha for each pair of samples)

$Pairwise_samples$Gamma (gamma for each pair of samples)

$Pairwise_samples$Beta_add (beta for each pair of samples as Gamma-Mean_Alpha)

$Pairwise_samples$Beta_prop (beta for each pair of samples as Beta_add*100/Gamma)

References

De Bello, Francesco, Sandra Lavorel, Cecile H. Albert, Wilfried Thuiller, Karl Grigulis, Jiri Dolezal, stepan Janecek, et Jan Leps. 2011. Quantifying the relevance of intraspecific trait variability for functional diversity: Intraspecific variability in functional diversity. Methods in Ecology and Evolution 2: 163-174.

Author

Francesco De Bello et al., 2011 modified by Adrien Taudiere

Examples