cati: Community Assembly by Traits: Individuals and Beyond

Detect and quantify community assembly processes using trait values of individuals or populations, the T-statistics and other metrics, and dedicated null models. Provides tools to analyse 'intraspecific' trait variability and its consequences for community assembly. Implements a framework using individual-level trait data to decompose variance at the population, species, and community levels. Methods are described in 'Taudiere' and 'Violle' (2016) doi:10.1111/ecog.01433 .

See also

Author

Maintainer: Adrien Taudiere adrien.taudiere@zaclys.net (ORCID) [copyright holder]

Authors:

Cyrille Violle (ORCID)

Other contributors:

François Munoz (ORCID) [contributor]