Plot result of observed indices values against null distribution

Function to plot result of observed indices values against null distribution.

Usage

plotRandtest(x, alternative = "two-sided", ...)

Arguments

x

An object of class listofindex, ComIndex, ComIndexMulti or Tstats.

alternative

Indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less". You can specify just the initial letter. "greater" corresponds to positive association, "less" to negative association.

...

Any additional arguments are passed to the plot function creating the core of the plot and can be used to adjust the look of resulting graph.

Value

None; used for the side-effect of producing a plot.

Author

Adrien Taudiere

See also

Examples

  data(finch.ind)
  if (FALSE) { # \dontrun{
    res.finch <- Tstats(traits.finch, ind.plot = ind.plot.finch, 
    sp = sp.finch, nperm = 99, print = FALSE)
  
  par(mfrow = c(4,4))
  
  plotRandtest(res.finch)
  plotRandtest(res.finch, alter = "less")
  } # }