Standardized effect size and confidence interval for a matrix of statistics

calculation standardized effect size and confidence interval for a matrix of statistics and the related null model expressed as a list or as an array. Internal function use by other functions of the package. You can transpose the observed matrix to represent either the SES by traits or by plots.

Usage

ses(obs = NULL, nullmodel = NULL, val.quant = c(0.025, 0.975))

Arguments

obs

Observed matrix or vector of values.

nullmodel

Either a list or an array of three (two for a vector of observed values) dimensions corresponding to the null model permutations.

val.quant

Numeric vectors of length 2, giving the quantile to calculation confidence interval. By default val.quant = c(0.025,0.975) for a bilateral test with alpha = 5%.

Details

Warning: to detect automatically the correspondence between dimension of observed matrix and null model list or array, observed matrix needs to have different numbers of rows and columns. In the case of same row and column number, please verify manually the correspondance beatween the rows of the observed matrix and the null model array.

Value

A list of three components:

$ses

Observed value of standardized effect size.

$ses.inf

Lower limit of the confidence interval.

$ses.sup

Upper limit of the confidence interval.

Author

Adrien Taudiere

See also

Examples

  data(finch.ind)
  
  res.finch <- Tstats(traits.finch, ind.plot = ind.plot.finch, 
  sp = sp.finch, nperm = 9)
#> Warning: This function exclude 1137 Na values
#> [1] "creating null models"
#> [1] "8.33 %"
#> [1] "16.67 %"
#> [1] "25 %"
#> [1] "33.33 %"
#> [1] "41.63 %"
#> [1] "49.97 %"
#> [1] "58.3 %"
#> [1] "66.63 %"
#> [1] "74.93 %"
#> [1] "83.27 %"
#> [1] "91.6 %"
#> [1] "99.93 %"
#> [1] "calculation of Tstats using null models"
#> [1] "8.33 %"
#> [1] "16.67 %"
#> [1] "25 %"
#> [1] "33.33 %"
#> [1] "41.63 %"
#> [1] "49.97 %"
#> [1] "58.3 %"
#> [1] "66.63 %"
#> [1] "74.93 %"
#> [1] "83.27 %"
#> [1] "91.6 %"
#> [1] "99.93 %"
  if (FALSE) { # \dontrun{
    ses(res.finch$Tstats$T_IP.IC, res.finch$Tstats$T_IP.IC_nm)
  } # }