Standardized effect size for a list of index.

Standardized effect size and confidence interval for a list of index.

Usage

ses.listofindex(index.list = NULL, val.quant = c(0.025, 0.975))

Arguments

index.list

A list of index obtain using the function as.listofindex.

val.quant

Numeric vectors of length 2, giving the quantile to calculation confidence interval. By default val.quant = c(0.025,0.975) for a bilateral test with alpha = 5%.

Value

A list which each component correspond to the result of the ses function for an index. Further, each component is a list of three components:

$ses

Observed value of standardized effect size.

$ses.inf

Lower limit of the confidence interval.

$ses.sup

Upper limit of the confidence interval.

Author

Adrien Taudiere

See also